add streaming

+++ src/helpers/helper.py
@@ -40,7 +40,7 @@ def determine_block_lat_ids(lat):
     :rtype : (int, int)       tuples which contain x and y ids
     """
     # size of large block is 0.005  degree lat/lon, about 350 meters
-    corner = lat - 35.98
+    corner = float(lat) - 35.98
     block_id_lat = int(math.floor(corner/0.005))
     return block_id_lat
 
@@ -53,7 +53,7 @@ def determine_block_log_ids(log):
     :rtype : (int, int)       tuples which contain x and y ids
     """
     # size of large block is 0.005  degree lat/lon, about 350 meters
-    corner = log + 115.65
+    corner = float(log) + 115.65
     block_id_log = int(math.floor(corner/0.005))
     return block_id_log
